+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export { AddEventModal } from './AddEventModal';
 export { CompareFilters } from './CompareFilters';
 export { CompareModal } from './CompareModal';
+export { EditVenueModal } from './EditVenueModal';
 export { EventsCompareView } from './EventsCompareView';
 export { EventsList } from './EventsList';
 export { ExportMenu } from './ExportMenu';